+++ tgsw.h
@@ -14,15 +14,15 @@ void tgsw_sample_clear(tgsw_sample ct);
 void tgsw_sample_zero(tgsw_sample *ct);
 
 /* Encrypts polynomial m using key sk and noise parameter param */
-tgsw_sample tgsw_encrypt(tlwe_sk sk, long double param, int* m);
-void tgsw_encrypt_over(tgsw_sample ct, tlwe_sk sk, long double param, int* m);
+tgsw_sample tgsw_encrypt(tlwe_sk sk, double param, int* m);
+void tgsw_encrypt_over(tgsw_sample ct, tlwe_sk sk, double param, int* m);
 
 /* Decrypts ciphertext ct using key sk */
 int* tgsw_decrypt(tlwe_sk sk, tgsw_sample ct);
 void tgsw_decrypt_over_and_keep(int *m, tlwe_sk sk, tgsw_sample ct);
 
 /* Generates a boostrapping key */
-bsk generate_bsk(lwe_sk sk_in, tlwe_sk sk_out, long double param, size_t n);
+bsk generate_bsk(lwe_sk sk_in, tlwe_sk sk_out, double param, size_t n);
 void bsk_clear(bsk bsk, size_t n);
 
 /* Exeternal product between a TLWE sample and a TGSW sample */
