+++ tlwe.h
@@ -32,10 +32,10 @@ void multiply_accumulate_poly(uint64_t *out, uint64_t *a, uint64_t *b);
 void decompose_tlwe(uint64_t *out, tlwe_sample in);
 
 /* Encrypts polynomial m using key sk, message space M and noise parameter/stdev param */
-void tlwe_encrypt_over(tlwe_sample ct, tlwe_sk sk, long double param, int M, int* m);
-tlwe_sample tlwe_encrypt(tlwe_sk sk, long double param, int M, int* m);
-void tlwe_encrypt_zero_over(tlwe_sample ct, tlwe_sk sk, long double param);
-tlwe_sample tlwe_encrypt_zero(tlwe_sk sk, long double param);
+void tlwe_encrypt_over(tlwe_sample ct, tlwe_sk sk, double param, int M, int* m);
+tlwe_sample tlwe_encrypt(tlwe_sk sk, double param, int M, int* m);
+void tlwe_encrypt_zero_over(tlwe_sample ct, tlwe_sk sk, double param);
+tlwe_sample tlwe_encrypt_zero(tlwe_sk sk, double param);
 void tlwe_encrypt_zero_gaussian_over(tlwe_sample ct, tlwe_sk sk, gaussian_param_t param);
 tlwe_sample tlwe_encrypt_zero_gaussian(tlwe_sk sk, gaussian_param_t param);
 
@@ -49,10 +49,9 @@ lwe_sk tlwe_key_extract(tlwe_sk in);
 lwe_sample tlwe_extract(tlwe_sample in);
 void tlwe_extract_over_and_keep(lwe_sample out, tlwe_sample in);
 
-
 /* Generates public keys necessary for sanitization of ciphertexts */
 // No longer required -> pkc is fft and pks is lwe
-//pkg sanitize_pkc_gen(tlwe_sk tsk, long double param);
+//pkg sanitize_pkc_gen(tlwe_sk tsk, double param);
 //void sanitize_pkc_clear(pkg PK);
 
 /* Generates a well distributed TLWE encryption of 0 and adds it to the ciphertext */
